Extract useCollapse into hook
